+++ NEWS.md
@@ -1,7 +1,7 @@
 # teal.goshawk 0.1.8.9000
 
 ### Enhancements
-* Updated `tm_g_gh_correlationplot`, `tm_g_gh_scatterplot` and `tm_g_gh_boxplot` encodings to have a checkbox to facet by the treatment variable instead of 
+* Updated `tm_g_gh_correlationplot` and `tm_g_gh_scatterplot` encodings to have a checkbox to facet by the treatment variable instead of 
 a drop down menu.
 
 # teal.goshawk 0.1.8